Move MovieData type to seperate file
